+++ config.json
@@ -1,1 +1,1 @@
-multi_column_editor/config.json
+{"MAX_COLUMNS": 18}
