+++ editor.js
