--- multi_column_editor/config.json
@@ -1,1 +1,0 @@
-{"MAX_COLUMNS": 18}
